drivevlm_nuscenes/requirements.txt
nuscenes-devkit
opencv-python
openai
pillow
numpy
gtts
pydub
moviepy
scipy
